feat:Item Entity confirm #24

+++ src/main/java/kr/co/d_erp/dtos/CatDto.java
@@ -1,6 +1,7 @@
 package kr.co.d_erp.dtos;
 
 import jakarta.persistence.Column;
+
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
@@ -14,11 +15,11 @@ import lombok.Setter;
 public class CatDto {
     @Id
     @Column(name = "CAT_IDX")
-    private Integer catIdx;
+    private Long catIdx;
 
     @Column(name = "CAT_NM")
     private String catNm;
 
     @Column(name = "PARENT_IDX")
-    private Integer parentIdx; // 대분류의 경우 null 또는 0
+    private Long parentIdx; // 대분류의 경우 null 또는 0
 }
